Update to 2.3.3 for client.

+++ src/main/java/pl/arturekdev/mineEconomy/comparator/RankingSort.java
@@ -1,7 +1,7 @@
 package pl.arturekdev.mineEconomy.comparator;
 
 
-import pl.arturekdev.mineEconomy.objects.User;
+import pl.arturekdev.mineEconomy.user.User;
 
 import java.util.Comparator;
 
@@ -9,6 +9,6 @@ public class RankingSort implements Comparator<User> {
 
     @Override
     public int compare(User u1, User u2) {
-        return u2.getMoney() - u1.getMoney();
+        return (int) (u2.getMoney() - u1.getMoney());
     }
 }
